Knekt kodeverk opp i flere filer
