+++ src/main/kotlin/no/nav/k9/los/kodeverk/BehandlingStatus.kt
@@ -1,0 +1,30 @@
+package no.nav.k9.los.kodeverk
+
+import com.fasterxml.jackson.annotation.JsonCreator
+import com.fasterxml.jackson.annotation.JsonFormat
+
+@JsonFormat(shape = JsonFormat.Shape.OBJECT)
+enum class BehandlingStatus(override val kode: String, override val navn: String) : Kodeverdi {
+    AVSLUTTET("AVSLU", "Avsluttet"),
+    FATTER_VEDTAK("FVED", "Fatter vedtak"),
+    IVERKSETTER_VEDTAK("IVED", "Iverksetter vedtak"),
+    OPPRETTET("OPPRE", "Opprettet"),
+    UTREDES("UTRED", "Utredes"),
+
+    // de 3 siste gjelder k9-punsj
+    SATT_PÅ_VENT("VENT", "Satt på vent"),
+    LUKKET("LUKKET", "Lukket"),
+    SENDT_INN("SENDT_INN", "Sendt inn");
+
+    override val kodeverk = "BEHANDLING_TYPE"
+
+    companion object {
+        @JsonCreator(mode = JsonCreator.Mode.DELEGATING)
+        @JvmStatic
+        fun fraKode(o: Any): BehandlingStatus {
+            val kode = TempAvledeKode.getVerdi(o)
+            return values().find { it.kode == kode } ?: throw IllegalStateException("Kjenner ikke igjen koden=$kode")
+        }
+    }
+}
+
