+++ src/main/kotlin/no/nav/k9/los/kodeverk/BehandlingType.kt
@@ -1,0 +1,49 @@
+package no.nav.k9.los.kodeverk
+
+import com.fasterxml.jackson.annotation.JsonCreator
+import com.fasterxml.jackson.annotation.JsonFormat
+
+@JsonFormat(shape = JsonFormat.Shape.OBJECT)
+enum class BehandlingType(override val kode: String, override val navn: String, override val kodeverk: String) :
+    Kodeverdi {
+    FORSTEGANGSSOKNAD("BT-002", "Førstegangsbehandling", "ae0034"),
+    KLAGE("BT-003", "Klage", "ae0058"),
+    REVURDERING("BT-004", "Revurdering", "ae0028"),
+    INNSYN("BT-006", "Innsyn", "ae0042"),
+    TILBAKE("BT-007", "Tilbakekreving", "ae0203"),
+    ANKE("BT-008", "Anke", "ae0046"),
+    REVURDERING_TILBAKEKREVING("BT-009", "Tilbakekreving revurdering", "BT-009"),
+    UNNTAKSBEHANDLING("BT-010", "Unntaksbehandling", "N/A"),
+    SØKNAD_OM_NYE_PERIODER("BT-011", "Søknad om nye perioder", "N/A"),
+
+    //gjelder punsj
+    PAPIRSØKNAD("PAPIRSØKNAD", "Papirsøknad", "PUNSJ_INNSENDING_TYPE"),
+    DIGITAL_SØKNAD("DIGITAL_SØKNAD", "Digital søknad", "PUNSJ_INNSENDING_TYPE"),
+    PAPIRETTERSENDELSE("PAPIRETTERSENDELSE", "Papirettersendelse", "PUNSJ_INNSENDING_TYPE"),
+    PAPIRINNTEKTSOPPLYSNINGER("PAPIRINNTEKTSOPPLYSNINGER", "Papirinntektsopplysninger", "PUNSJ_INNSENDING_TYPE"),
+    DIGITAL_ETTERSENDELSE("DIGITAL_ETTERSENDELSE", "Digital ettersendelse", "PUNSJ_INNSENDING_TYPE"),
+    INNLOGGET_CHAT("INNLOGGET_CHAT", "Innlogget chat", "PUNSJ_INNSENDING_TYPE"),
+    SKRIV_TIL_OSS_SPØRMSÅL("SKRIV_TIL_OSS_SPØRMSÅL", "Skriv til oss spørsmål", "PUNSJ_INNSENDING_TYPE"),
+    SKRIV_TIL_OSS_SVAR("SKRIV_TIL_OSS_SVAR", "Skriv til oss svar", "PUNSJ_INNSENDING_TYPE"),
+    SAMTALEREFERAT("SAMTALEREFERAT", "Samtalereferat", "PUNSJ_INNSENDING_TYPE"),
+    KOPI("KOPI", "Kopi", "PUNSJ_INNSENDING_TYPE"),
+    INNTEKTSMELDING_UTGÅTT("INNTEKTSMELDING_UTGÅTT", "Inntektsmeldinger uten søknad", "PUNSJ_INNSENDING_TYPE"),
+    UTEN_FNR_DNR("UTEN_FNR_DNR", "Uten fnr eller dnr", "PUNSJ_INNSENDING_TYPE"),
+    PUNSJOPPGAVE_IKKE_LENGER_NØDVENDIG("PUNSJOPPGAVE_IKKE_LENGER_NØDVENDIG", "Punsjoppgave ikke lenger nødvendig", "PUNSJ_INNSENDING_TYPE"), // Prodfiks: Lagt til fordi den er mottatt på kafka
+    JOURNALPOSTNOTAT("JOURNALPOSTNOTAT", "Manuelt opprettet journalpostnotat", "PUNSJ_INNSENDING_TYPE"),
+    UKJENT("UKJENT", "Ukjent", "PUNSJ_INNSENDING_TYPE");
+
+    companion object {
+        @JsonCreator(mode = JsonCreator.Mode.DELEGATING)
+        @JvmStatic
+        fun fraKode(o: Any): BehandlingType {
+            val kode = TempAvledeKode.getVerdi(o)
+            return entries.find { it.kode == kode } ?: throw IllegalStateException("Kjenner ikke igjen koden=$kode")
+        }
+    }
+
+    fun gjelderPunsj() : Boolean {
+        return kodeverk == "PUNSJ_INNSENDING_TYPE"
+    }
+}
+
