+++ src/main/kotlin/no/nav/k9/los/kodeverk/FagsakYtelseType.kt
@@ -1,0 +1,31 @@
+package no.nav.k9.los.kodeverk
+
+import com.fasterxml.jackson.annotation.JsonCreator
+import com.fasterxml.jackson.annotation.JsonFormat
+
+@JsonFormat(shape = JsonFormat.Shape.OBJECT)
+enum class FagsakYtelseType constructor(override val kode: String, override val navn: String) : Kodeverdi {
+    PLEIEPENGER_SYKT_BARN("PSB", "Pleiepenger sykt barn"),
+    OMSORGSPENGER("OMP", "Omsorgspenger"),
+    OMSORGSDAGER("OMD", "Omsorgsdager: overføring"),
+    FRISINN("FRISINN", "Frisinn"),
+    PPN("PPN", "Pleiepenger i livets sluttfase"),
+    OLP("OLP", "Opplæringspenger"),
+    OMSORGSPENGER_KS("OMP_KS", "Omsorgsdager: kronisk syk"),
+    OMSORGSPENGER_MA("OMP_MA", "Omsorgsdager: midlertidig alene"),
+    OMSORGSPENGER_AO("OMP_AO", "Omsorgsdager: alene om omsorg"),
+    UNGDOMSYTELSE("UNG", "Ungdomsytelse"),
+    UKJENT("UKJENT", "Ukjent");
+
+    override val kodeverk = "FAGSAK_YTELSE_TYPE"
+
+    companion object {
+        @JsonCreator(mode = JsonCreator.Mode.DELEGATING)
+        @JvmStatic
+        fun fraKode(o: Any): FagsakYtelseType {
+            val kode = TempAvledeKode.getVerdi(o)
+            return values().find { it.kode == kode } ?: throw IllegalStateException("Kjenner ikke igjen koden=$kode")
+        }
+    }
+}
+
